Increase nigiri time window to one year
This does not seem to have any adverse effects on the memory consumption
or query performance on my local Switzerland example at least, but it's
significantly more useful.

Going beyond one year brings little benefit as the GTFS schedules don't
seem to be valid that far into the future anyway.