website: Add license page
I ended up changing the format of license.json a bit, but nobody else should
be using it currently.

For feeds were we require a mirror, the original url is now kept to link
to it on the license page (It's also useful to document were the mirror
pulls from).