Add more feeds in Romania
I added an escape hatch to not use the mirrored version from Mobility
Database, because sometimes they are blocked as well.

I'm not entirely happy about the different semantics of proxy for
transitland and use-origin, but neither do I want to turn proxy on
unconditionally nor loose the advantage of the mobilitydatabase cache.