Add support for MOTIS v2.2 import scripts
The idea here is to have a top-level folder "scripts" containing
scripts named after the feeds they are for.

The upside of this is that we don't need to maintain a separate
configuration, and common code needed for multiple feeds can be
factored out into a shared file.

The downside is that any deviation in the naming will make this
silently ignore the script.